core of analitic service

+++ CaseProfile/internal/case_profile/models/case_profile.go
@@ -19,12 +19,12 @@ type CaseProfile struct {
 	ID                   int64     `json:"id" db:"id"`
 	UserID               int64     `json:"user_id" db:"user_id"`
 	TotalCases           int64     `json:"total_cases" db:"total_cases"`
-	Assertiveness        float64   `json:"assertiveness" db:"assertiveness"`
-	Empathy              float64   `json:"empathy" db:"empathy"`
-	ClarityCommunication float64   `json:"clarity_communication" db:"clarity_communication"`
-	Resistance           float64   `json:"resistance" db:"resistance"`
-	Eloquence            float64   `json:"eloquence" db:"eloquence"`
-	Initiative           float64   `json:"initiative" db:"initiative"`
+	Assertiveness        float32   `json:"assertiveness" db:"assertiveness"`
+	Empathy              float32   `json:"empathy" db:"empathy"`
+	ClarityCommunication float32   `json:"clarity_communication" db:"clarity_communication"`
+	Resistance           float32   `json:"resistance" db:"resistance"`
+	Eloquence            float32   `json:"eloquence" db:"eloquence"`
+	Initiative           float32   `json:"initiative" db:"initiative"`
 	ChangedAt            time.Time `json:"changed_at" db:"changed_at"`
 }
